Switched around the file structure of project.
