+++ structures/structures.go
@@ -1,0 +1,7 @@
+package structures
+
+type Tree struct {
+	Board    [8][8]string
+	Children []*Tree // we will use append and make with this slice
+	// to check if it is a leaf, we can check if Children is nil
+}
